Live Relay: in-memory receive store (last position per sender) for the map
Replace LiveLocationDebugLogger with LiveLocationReceiveStore: an in-memory
StateFlow<Map<OdinId, LivePosition>> of each sender's last known position +
server receipt time, fed from BackendEvent.LiveRelayReceived (last-value-wins),
cleared on logout. This is the data source a future map/dashboard binds to; it
still logs RECV-DECODED.

Deliberately NOT persisted: the server retains each sender's last point and
auto-flushes on (re)connect/foreground, so the map rehydrates from the server on
a cold start. Persisting would duplicate that and risk rendering positions
staler than the server's ~5min TTL. (Asymmetry with the send roster, which IS
persisted because it's the user's own intent.) Plan doc updated with a hydration
note + the in-memory-only rationale.
